add S3 to WardrobeService + test services update & storage

+++ backend/app/services/wardrobe_service.py
@@ -1,6 +1,6 @@
 from app.repositories.wardrobe_repo import WardrobeRepository
-from app.repositories.storage import StorageRepository
-from app.api.schemas.wardrobe import ClothResponse, ClothCreate, ClothCreateResponse, ClothListResponse, ClothDeleteResponse
+from app.repositories.storage_repo import StorageRepository
+from app.api.schemas.wardrobe_schema import ClothResponse, ClothCreate, ClothCreateResponse, ClothListResponse, ClothDeleteResponse
 from app.core.errors import NotFoundError
 from app.core.logging_config import logger
 from datetime import datetime
@@ -9,7 +9,7 @@ import uuid
 class WardrobeService:
     def __init__(self, repository: WardrobeRepository = None, storage_repo: StorageRepository = None):
         self.repository = repository or WardrobeRepository()
-        self.storage_repo = StorageRepository()
+        self.storage_repo = storage_repo or StorageRepository()
 
     async def create_cloth(self, cloth: ClothCreate):
         """
@@ -21,35 +21,49 @@ class WardrobeService:
         """
         logger.info(f"🟡 [Service] Creating new cloth in repository")
 
-        # Create a unique ID for the cloth
+        # Generate unique ID for the cloth
         cloth_id = str(uuid.uuid4())
 
-        # Upload image to S3 TODO: Fix S3
-        #image_url = await self.storage_repo.upload_cloth_image(cloth.user_id, cloth_id, str(cloth.image_url))
-        image_url = str(cloth.image_url)
-        #if not image_url:
-        #    logger.error(f"🔴 [Service] Failed to upload image to S3")
-        #    raise Exception("Failed to upload image to S3")
+        # Upload original image to S3
+        image_url = await self.storage_repo.upload_cloth_image(cloth.user_id, cloth_id, cloth.file)
 
-        # Prepare cloth object
+        if not image_url:
+            logger.error(f"🔴 [Service] Failed to upload image to S3")
+            raise Exception("Failed to upload image to S3")
+        
+        # TODO: Call pre-processing service for cloth
+        # Call pre-processing service for cloth
+        #cloth_preprocessed = await self.preprocess_service.process_cloth(cloth.file)
+
+        # Save cloth_preprocessed to S3
+        #image_url_preprocessed = await self.storage_repo.upload_cloth_image(cloth.user_id, cloth_id, cloth_preprocessed)
+
+        #if not image_url_preprocessed:
+        #    logger.error(f"🔴 [Service] Failed to upload preprocessed image to S3")
+        #    raise Exception("Failed to upload preprocessed image to S3")
+
+        # Save cloth to database
         data = {
+            "_id": cloth_id,
             "user_id": cloth.user_id,
             "name": cloth.name,
             "type": cloth.type,
-            "image_url": str(image_url),
+            "image_url": image_url,
             "created_at": datetime.now()
         }
 
-        cloth_id = await self.repository.create_cloth(data)
+        inserted_id = await self.repository.create_cloth(data)
 
-        if not cloth_id:
-            logger.error(f"🔴 [Service] Failed to create cloth")
+        if not inserted_id:
+            logger.error(f"🔴 [Service] Failed to create cloth in database")
             raise Exception("Failed to create cloth")
+        
         
         logger.debug(f"🟢 [Service] Cloth created with ID: {cloth_id}")
         return ClothCreateResponse(
             id=str(cloth_id), 
             message=f"Cloth created successfully",
+            image_url=image_url,
             created_at=data["created_at"]
         )
 
@@ -90,16 +104,28 @@ class WardrobeService:
     
     async def delete_cloth(self, cloth_id: str):
         logger.info(f"🟡 [Service] Deleting cloth {cloth_id} from repository")
+
+        # Get cloth from database
         cloth = await self.repository.get_cloth_by_id(cloth_id)
 
         if not cloth:
             logger.warning(f"🔴 [Service] Cloth {cloth_id} not found")
             raise NotFoundError(f"Cloth {cloth_id} not found")
 
-        # Delete image from S3 TODO: Fix S3
-        #await self.storage_repo.delete_cloth_image(cloth["user_id"], cloth_id)
-
-        await self.repository.delete_cloth(cloth_id)
+        # Delete image from S3
+        success_s3 = await self.storage_repo.delete_cloth_image(cloth["user_id"], cloth_id)
+        
+        if not success_s3:
+            logger.error(f"🔴 [Service] Failed to delete image from S3")
+            raise Exception("Failed to delete image from S3")
+        
+        # Delete cloth from database
+        succes_db = await self.repository.delete_cloth(cloth_id)
+        
+        if not succes_db:
+            logger.error(f"🔴 [Service] Failed to delete cloth from database")
+            raise Exception("Failed to delete cloth from database")
+        
         logger.debug(f"🟢 [Service] Cloth {cloth_id} deleted")
         return ClothDeleteResponse(
             message=f"Cloth {cloth_id} deleted successfully"
